requirements.txt
aiogram>=3.0.0
requests>=2.28.1
beautifulsoup4>=4.11.1
python-dotenv>=1.0.0
aiohttp>=3.8.1
pandas>=1.5.0
openpyxl>=3.0.10
lxml>=4.9.1
rich>=12.6.0